pyproject.toml
[tool.poetry]
name = "fastapi-pypoetry-template"
version = "0.1.0"
description = ""
authors = ["BruceMcD <162571835+BruceMcD@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
numpy = "^2.0.0"
sqlalchemy = "^2.0.31"
aiomysql = "^0.2.0"
mysql-connector-python = "^9.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

